+++ API/src/main/java/fr/maxlego08/essentials/api/worldedit/WorldEditTask.java
@@ -402,6 +402,9 @@ public abstract class WorldEditTask {
             }
 
             block.setType(blockInfo.newMaterial());
+            if (!blockInfo.newMaterial().isAir()) {
+                plugin.getBlockTracker().track(block);
+            }
         });
     }